backend/requirements.txt
# Django Core
Django==5.1.3
django-cors-headers==4.6.0
python-decouple==3.8

# Database
psycopg[binary]==3.2.3
django-extensions==3.2.3

# API
django-ninja==1.3.0
pydantic==2.10.3
email-validator==2.1.0

# Authentication
PyJWT==2.10.1
django-allauth==65.3.0

# File Storage
pillow==11.0.0
boto3==1.35.81
django-storages==1.14.4

# Development
ipython==8.30.0
django-debug-toolbar==4.4.6

# Testing
pytest==8.3.4
pytest-django==4.9.0
pytest-cov==6.0.0
factory-boy==3.3.1

# Utilities
python-slugify==8.0.4
django-filter==24.3
django-guardian==2.4.0

# Geolocation
geopy==2.4.1

# Web Scraping
beautifulsoup4==4.12.3
requests==2.32.3
lxml==5.3.0
selenium==4.27.1  # Optional, for JavaScript-heavy sites

# Task Queue (for scheduled crawling)
celery==5.4.0
redis==5.2.1
django-celery-beat==2.7.0

# Static Files
whitenoise==6.7.0